Finalized app logic. Now only pre excel export calculations remain.

+++ lib/controllers/properties/all_properties_controller.dart
@@ -3,10 +3,12 @@ import 'package:get/get.dart';
 import 'package:isar/isar.dart';
 import 'package:real_estate_allotment/controllers/studies/active_study_controller.dart';
 import 'package:real_estate_allotment/core/services/isar_service.dart';
+import 'package:real_estate_allotment/models/allotment/lot_allotment/lot_allotment.dart';
+import 'package:real_estate_allotment/models/allotment/real_estate_allotment/real_estate_allotment.dart';
 import 'package:real_estate_allotment/models/lot/lot.dart';
 import 'package:real_estate_allotment/models/real_estate/real_estate.dart';
 
-enum DeleteResult { success, lotError, error }
+enum DeleteResult { success, lotError, error, allotmentError }
 
 class AllPropertiesController extends GetxController {
   final isar = Get.find<IsarService>().isar;
@@ -31,6 +33,12 @@ class AllPropertiesController extends GetxController {
       final result = await isar.writeTxn(
         () async {
           bool success;
+
+          success = await _deleteAllotments(
+            propertyId: propertyId,
+          );
+          if (!success) return DeleteResult.allotmentError;
+
           success = await _deleteLots(
             propertyId: propertyId,
           );
@@ -50,18 +58,41 @@ class AllPropertiesController extends GetxController {
     }
   }
 
+  Future<bool> _deleteAllotments({required int propertyId}) async {
+    final allotmentIdsList = await isar.realEstateAllotments
+        .where()
+        .propertyIdEqualToAnyShareholderName(propertyId)
+        .idProperty()
+        .findAll();
+    final result = await isar.realEstateAllotments.deleteAll(allotmentIdsList);
+    return (result == allotmentIdsList.length);
+  }
+
   Future<bool> _deleteLots({required int propertyId}) async {
     final lotIdsList = await isar.lots
         .where()
         .propertyIdEqualToAnyLotNumber(propertyId)
         .idProperty()
         .findAll();
-    final result = await isar.lots.deleteAll(lotIdsList);
 
-    if (result == lotIdsList.length) {
-      return true;
-    } else {
-      return false;
+    bool success = false;
+    for (var lotId in lotIdsList) {
+      success = await _deleteLotAllotments(lotId: lotId);
+      if (!success) break;
     }
+    if (!success) return false;
+
+    final result = await isar.lots.deleteAll(lotIdsList);
+    return (result == lotIdsList.length);
+  }
+
+  Future<bool> _deleteLotAllotments({required int lotId}) async {
+    final allotmentIdsList = await isar.lotAllotments
+        .where()
+        .lotIdEqualToAnyShareholderName(lotId)
+        .idProperty()
+        .findAll();
+    final result = await isar.lotAllotments.deleteAll(allotmentIdsList);
+    return (result == allotmentIdsList.length);
   }
 }
